overhaul the working of "checkout"
the CApp checkout operation
(used during fresh cloning and
 also the CApp checkout command)
is now able to handle very complex
cases of multiple remotes and branches.
First, if the desired Git URL
is not a remote in the existing
local repository, then it will be
added and fetched as a remote
(which is much more efficient and
consistent with developer workflow).
Second, if the desired commit
corresponds to a branch in the
desired Git URL, then that branch
will be checked out for convenience.